fix(schemas): CheckoutSummaryResponse money fields are Decimal, not float (M8)
Companion to the M8 Decimal migration — subtotal/tax/grand_total feed
from Order.total_amount (already Decimal after the prior fix) and a
GST calculation in urjaa-backend that's being fixed alongside this.

+++ urjaa_core/schemas/checkout.py
@@ -82,8 +82,8 @@ class CheckoutSummaryRequest(BaseModel):
 
 class CheckoutSummaryResponse(BaseModel):
     order_id: UUID
-    subtotal: float
-    tax: float
-    grand_total: float
+    subtotal: Decimal
+    tax: Decimal
+    grand_total: Decimal
     formatted_tax: str
     formatted_grand_total: str
